prompt delete profile now actually shows a prompt

+++ app/src/main/java/ie/delilahsthings/soothingloop/SettingsActivity.java
@@ -1,5 +1,6 @@
 package ie.delilahsthings.soothingloop;
 
+import android.app.AlertDialog;
 import android.os.Bundle;
 import android.view.View;
 import android.view.ViewGroup;
@@ -40,7 +41,17 @@ public class SettingsActivity extends AppCompatActivity {
 
     void promptDeleteProfile(View sender, String profileName)
     {
-        ProfileManager.deleteProfile(this,profileName);
-        profileList.removeView(sender);
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setTitle(String.format(getString(R.string.confirm_delete_profile),profileName));
+        builder.setNegativeButton(R.string.cancel, (dialogInterface, i) -> dialogInterface.cancel());
+        builder.setPositiveButton(getString(R.string.confirm),(dialogInterface, i) -> {
+            ProfileManager.deleteProfile(this,profileName);
+            profileList.removeView(sender);
+        });
+
+        AlertDialog dialog = builder.create();
+        dialog.show();
+
+        Workarounds.fixBadDialogButtonColours(this, dialog);
     }
 }
